rate limit

+++ Src/Base/Persistence/GenericRepositoryExtensions.cs
@@ -6,7 +6,15 @@ using System.Linq;
 
 public static class GenericRepositoryExtensions
 {
-    public static void Sync<TEntity, TKey>(this ICollection<TEntity> inDb, ICollection<TEntity> toDb, Func<TEntity, TKey> keySelector, Action<TEntity, TEntity>? copyToDb = null, Action<TEntity>? setAdd = null, Action<TEntity>? setDelete = null)
+    public static void Sync<TEntity, TKey>(
+        this ICollection<TEntity> inDb,
+        ICollection<TEntity>      toDb,
+        Func<TEntity, TKey>       keySelector,
+        Action<TEntity, TEntity>? copyToDb   = null,
+        Action<TEntity>?          setAdd     = null,
+        Action<TEntity>?          setDelete  = null,
+        bool                      skipAdd    = false,
+        bool                      skipDelete = false)
     {
         foreach (var ed in inDb.Join(toDb, keySelector, keySelector, (eDb, tDb) => (eDb, tDb)))
         {
@@ -16,16 +24,22 @@ public static class GenericRepositoryExtensions
         var toAdd = toDb.ExceptBy(inDb.Select(keySelector), keySelector).ToList();
         var toDel = inDb.ExceptBy(toDb.Select(keySelector), keySelector).ToList();
 
-        foreach (var add in toAdd)
+        if (!skipAdd)
         {
-            inDb.Add(add);
-            setAdd?.Invoke(add);
+            foreach (var add in toAdd)
+            {
+                inDb.Add(add);
+                setAdd?.Invoke(add);
+            }
         }
 
-        foreach (var del in toDel)
+        if (!skipDelete)
         {
-            setDelete?.Invoke(del);
-            inDb.Remove(del);
+            foreach (var del in toDel)
+            {
+                setDelete?.Invoke(del);
+                inDb.Remove(del);
+            }
         }
     }
 }
